--- zapdados/src/main/java/br/com/services/UsuarioService.java
@@ -1,16 +1,0 @@
-package br.com.services;
-
-import br.com.integration.UsuarioRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-/**
- *
- * @author thiagoespinhara
- */
-@Service
-public class UsuarioService {
-    
-    //@Autowired
-    //private UsuarioRepository repository;
-}
