+++ zapdados/src/main/java/br/com/zapdados/controllers/UrlListingListener.java
@@ -1,5 +1,7 @@
 package br.com.zapdados.controllers;
 
+import br.com.zapdados.integration.UsuarioRepository;
+import br.com.zapdados.model.Usuario;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.ApplicationListener;
@@ -13,14 +15,26 @@ public class UrlListingListener implements ApplicationListener<ApplicationReadyE
 
     @Autowired
     private RequestMappingHandlerMapping requestMappingHandlerMapping;
+    
+    @Autowired
+    private UsuarioRepository repository;
 
     @Override
     public void onApplicationEvent(ApplicationReadyEvent event) {
+        
+        addUsers();
+        
         Map<?, ?> handlerMethods = requestMappingHandlerMapping.getHandlerMethods();
         System.out.println("URLs disponíveis na aplicação:");
 
         handlerMethods.forEach((key, value) -> {
             System.out.println(key + " -> " + value);
         });
+    }
+    
+    private void addUsers() {
+        if (!repository.existsById("admin")) {
+            repository.save(new Usuario("admin", "admin"));
+        }
     }
 }
