+++ zapdados/src/main/java/br/com/zapdados/model/RelatorioExpressoes.java
@@ -1,4 +1,7 @@
-package br.com.services.model;
+package br.com.zapdados.model;
+
+import java.util.Map;
+import java.util.List;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,7 +9,6 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import java.util.List;
 import lombok.ToString;
 
 @Getter
@@ -16,11 +18,8 @@ import lombok.ToString;
 @AllArgsConstructor
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @ToString
-public class TempoUsoDados {
-    private int horaDoDia;
-    private String diaSemana;
-    private int dia;
-    private int mes;
-    private int ano;
-    private List<QtdUso> qtdUso;
+public class RelatorioExpressoes {
+     private String username;
+     private Map<String, Integer> expressoes;
+
 }
