+++ zapdados/src/main/java/br/com/zapdados/service/IUsuarioService.java
@@ -1,0 +1,9 @@
+package br.com.zapdados.service;
+
+/**
+ *
+ * @author thiagoespinhara
+ */
+public interface IUsuarioService {
+    boolean validarLogin(String user, String pass);
+}
