+++ zapdados/src/main/resources/application.properties
@@ -2,3 +2,16 @@ spring.application.name=zapdados
 
 # Habilitar a atualiza\u00e7\u00e3o autom\u00e1tica
 spring.devtools.livereload.enabled=true
+
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+
+# Configura\u00e7\u00f5es do H2
+spring.h2.console.enabled=true
+spring.datasource.url=jdbc:h2:file:/Users/thiagoespinhara/bd-h2/mydatabase
+spring.datasource.driver-class-name=org.h2.Driver
+spring.datasource.username=admin
+spring.datasource.password=admin
+
+# Configura\u00e7\u00f5es do JPA
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
